refactor: 리뷰사항에 따른 변경 적용
